move `HybridSet` to own file
